+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 > _Important: Please clear your browser cache (e.g. **Ctrl+F5**) to ensure outdated files are updated._
 
-- FIXED: `DNS` Tag is missing from the inbounds list in the routing rules form.
+- FIXED: Added missing `DNS` tag to the inbounds list in the routing rules form to ensure proper configuration.
 
 ## [0.42.9] - 2025-04-09